requirements.txt
auto-gptq==0.5.0.dev0
triton==2.1.0
python-dotenv==1.0.0
